Added: fixed a typo.

+++ Antenna/GMS_dipole_ant_types.f90
@@ -94,10 +94,10 @@ module dipole_ant_types
                                ! very thin dipole (3.4)
           integer(kind=i4) :: ndf34  ! number of dipoles in dipole array (3.4)
 #if (DIPOLE_ANT_TYPES_USE_PRECISION_REAL8) == 1
-          real(kind=sp), dimension(:,:), allocatable :: Iz
+          real(kind=dp), dimension(:,:), allocatable :: Iz
           !dir$ attributes align : 64 :: Iz
 #else
-          real(kind=dp), dimension(:,:), allocatable :: Iz
+          real(kind=sp), dimension(:,:), allocatable :: Iz
           !dir$ attributes align : 64 :: Iz 
 #endif
     end type Izf34
